#26 big refactor... god helps all
