+++ moviemancer/database_handlers.py
@@ -5,7 +5,6 @@ import hashlib
 import tmdbsimple as tmdb
 import time
 from decimal import Decimal, ROUND_HALF_UP
-from moviemancer.queries import *
 from moviemancer.tmdb_handler import TMDbHandler
 from moviemancer.helpers import Helpers
 
@@ -69,14 +68,83 @@ class DataBaseHandler:
         if not movie:
             DataBaseHandler.add_movie_to_database(tmdb_id=tmdb_movie_id)
             movie = Movie.objects.filter(tmdb_movie_id=tmdb_movie_id)
-            print('>>>>>>>>>>', movie[0].movie_id)
 
         DataBaseHandler.add_to_list(user_id=user_id, movie_id=movie[0].movie_id, list_type=list_type)
 
-    def rate_movie (user_id, movie_id, local_rate_id):
+    def rate_movie(user_id, movie_id, local_rate_id):
         DataBaseHandler.add_rating_to_movie(user_id, movie_id, local_rate_id)
 
         if Helpers.is_movie_on_list(user_id, movie_id, 2):
             DataBaseHandler.remove_movie_from_list(user_id, movie_id, 2)
 
         DataBaseHandler.add_to_list(user_id, movie_id, 3)
+
+    def rate_external_movie(user_id, user_rating, tmdb_movie_id, tmdb_poster, tmdb_title):
+        movie = Movie.objects.filter(tmdb_movie_id = tmdb_movie_id)
+
+        if movie:
+            movie_id = movie[0].movie_id
+            DataBaseHandler.rate_movie(user_id, movie_id, user_rating)
+            return True
+        else:
+            DataBaseHandler.add_movie_to_database(tmdb_id=tmdb_movie_id)
+            movie = Movie.objects.filter(tmdb_movie_id = tmdb_movie_id)
+            if movie:
+                movie_id = movie[0].movie_id
+                DataBaseHandler.rate_movie (user_id, movie_id, user_rating)
+                return True
+            return False
+
+    def create_user(name, email, password):
+        user = Viwer.objects.filter(email = email)
+
+        if not user:
+            hash_password = hashlib.sha224(password.encode('utf-8')).hexdigest()
+
+            user = Viwer(name=name, email=email, password=hash_password)
+            user.save()
+
+        return None
+
+    def create_list_to_user(user_id, type_id):
+        user_list = List.objects.filter(user_id = user_id, type_id = type_id)
+
+        if not user_list:
+            user_list = List(user_id = user_id, type_id = type_id)
+            user_list.save()
+
+    def register_user(name, email, password):
+        DataBaseHandler.create_user(name, email, password)
+
+        user_id = Helpers.get_user_id(email=email)
+
+        DataBaseHandler.create_list_to_user(user_id, 1)
+        DataBaseHandler.create_list_to_user(user_id, 2)
+        DataBaseHandler.create_list_to_user(user_id, 3)
+
+        return user_id
+
+    def update_user_info(user_id, email, name, password):
+        user = Helpers.handle_user_update_info(user_id, email, name, password)
+        user.save()
+
+    def authenticate_user(email, password):
+        hash_password = hashlib.sha224(password.encode('utf-8')).hexdigest()
+        user = Viwer.objects.filter(email=email, password=hash_password)
+
+        if user:
+            return True
+        return False
+
+    def add_comment(movie_tmdb_id, user_id, comment):
+        movie_id = Helpers.get_movie_id_by_tmdb_id(movie_tmdb_id)
+
+        comment = Comments(movie_id = movie_id, user_id = user_id, comment = comment)
+        comment.save()
+
+    def delete_comment(comment_id):
+        Comments.objects.get(comment_id = comment_id).delete()
+
+        if not Comments.objects.filter(comment_id = comment_id):
+            return True
+        return False
