+++ moviemancer/views.py
@@ -6,7 +6,7 @@ from moviemancer.models import Movie, Viwer
 from moviemancer.serializers import MovieSerializer, UserSerializer
 from rest_framework import generics
 from rest_framework.decorators import api_view
-from moviemancer.queries import *
+from moviemancer.helpers import Helpers
 from moviemancer.reco_refactor import generate_recommendation, update_recommendation
 from django.views.decorators.csrf import csrf_exempt
 from django.http import HttpResponse, HttpResponseServerError
@@ -68,7 +68,7 @@ def get_recommendation(request):
         request_reco = json.loads(request.body)
         request_user_id = request_reco[u'user_id']
 
-        recommendation_list = json.dumps(get_recommendation_list(request_user_id, 1))
+        recommendation_list = json.dumps(Helpers.get_recommendation_list(request_user_id, 1))
         return HttpResponse(recommendation_list)
     else:
         return HttpResponse("You are on your own")
@@ -93,7 +93,7 @@ def update_user(request):
         request_name = request_profile[u'name']
         request_password = request_profile[u'password']
 
-        update_user_info(request_user_id, request_email, request_name, request_password)
+        DataBaseHandler.update_user_info(request_user_id, request_email, request_name, request_password)
         return HttpResponse('Success')
     else:
         return HttpResponse("You are on your own")
@@ -118,7 +118,7 @@ def get_all_comments(request):
         request_comments = json.loads(request.body)
         request_movie_tmdb_id = request_comments[u'tmdb_movie_id']
 
-        all_comments = json.dumps(get_comments(request_movie_tmdb_id))
+        all_comments = json.dumps(Helpers.get_comments(request_movie_tmdb_id))
         return HttpResponse(all_comments)
     else:
         return HttpResponse("Erroa o carregar comentarios")
@@ -131,7 +131,7 @@ def add_new_comment(request):
         request_user_id = request_comments[u'user_id']
         request_comment = request_comments[u'comment']
 
-        add_comment(request_movie_tmdb_id, request_user_id, request_comment)
+        DataBaseHandler.add_comment(request_movie_tmdb_id, request_user_id, request_comment)
         return HttpResponse('Success')
     else:
         return HttpResponse("Erro ao criar comentario")
@@ -142,7 +142,7 @@ def delete_user_comment(request):
         request_comments = json.loads(request.body)
         request_comment_id = request_comments[u'comment_id']
 
-        if delete_comment(request_comment_id):
+        if DataBaseHandler.delete_comment(request_comment_id):
             return HttpResponse("Success")
         else:
             return HttpResponseServerError("Erro ao deletar comentario")
@@ -183,10 +183,10 @@ def get_auth(request):
         request_email = request_auth[u'email']
         request_password = request_auth[u'password']
 
-        if authenticate_user(request_email, request_password):
+        if DataBaseHandler.authenticate_user(request_email, request_password):
 
-            user_id = get_user_id_by_email(request_email)
-            user_data = json.dumps(get_user(request_email))
+            user_id = Helpers.get_user_id_by_email(request_email)
+            user_data = json.dumps(Helpers.get_user(request_email))
             generate_recommendation(user_id)
 
             return HttpResponse(user_data)
@@ -203,8 +203,8 @@ def registration(request):
         request_email = request_register[u'email']
         request_password = request_register[u'password']
 
-        if not user_exists(request_email):
-            user_id = register_user(request_name, request_email, request_password)
+        if not Helpers.user_exists(request_email):
+            user_id = DataBaseHandler.register_user(request_name, request_email, request_password)
             if user_id:
                 return HttpResponse(user_id)
             else:
@@ -218,7 +218,7 @@ def validate_user(request):
         request_validation = json.loads(request.body)
         request_email = request_validation[u'email']
 
-        if not user_exists(request_email):
+        if not Helpers.user_exists(request_email):
             return HttpResponse('Success')
         else:
             return HttpResponseServerError('User already registered')
@@ -233,7 +233,7 @@ def rate_external(request):
         request_movie_poster = request_user_rating[u'movie_poster']
         request_movie_title = request_user_rating[u'movie_title']
 
-        if rate_external_movie (request_user_id, request_rate_id, request_tmdb_movie_id, request_movie_poster, request_movie_title):
+        if DataBaseHandler.rate_external_movie (request_user_id, request_rate_id, request_tmdb_movie_id, request_movie_poster, request_movie_title):
             update_recommendation(request_user_id)
             return HttpResponse('Success')
         else:
@@ -275,7 +275,7 @@ def get_watched_list(request):
         request_user_rating = json.loads(request.body)
         request_user_id = request_user_rating[u'user_id']
 
-        user_watchedlist = json.dumps(get_watchedlist(request_user_id))
+        user_watchedlist = json.dumps(Helpers.get_watchedlist(request_user_id))
         return HttpResponse(user_watchedlist)
     else:
         return HttpResponse("You are on your own")
@@ -312,13 +312,13 @@ def get_watch_list(request):
         request_user_rating = json.loads(request.body)
         request_user_id = request_user_rating[u'user_id']
 
-        user_watchlist = json.dumps(get_watchlist(request_user_id))
+        user_watchlist = json.dumps(Helpers.get_watchlist(request_user_id))
         return HttpResponse(user_watchlist)
     else:
         return HttpResponse("You are on your own")
 
 def get_rated(request):
-    rated_movies = json.dumps(get_rated_movies())
+    rated_movies = json.dumps(Helpers.get_rated_movies())
     print(rated_movies)
     return HttpResponse(rated_movies)
 
@@ -334,7 +334,7 @@ def filter_reco(request):
         request_language = request_filter[u'language']
         request_genres = request_filter[u'genres']
 
-        filtered = json.dumps(filter_recommendation(request_genres, request_minYear, request_maxYear, request_minRuntime, request_maxRuntime, request_language, request_user_id))
+        filtered = json.dumps(Recommendation.filter_recommendation(request_genres, request_minYear, request_maxYear, request_minRuntime, request_maxRuntime, request_language, request_user_id))
         return HttpResponse(filtered)
     else:
         return HttpResponseServerError("You are on your own")
